fix(roads): cap circumcenter distance to reject near-degenerate Voronoi triangles
Near-collinear seed triples produce a near-zero determinant D, sending
the circumcenter to ±∞ while still passing the obtuse-triangle filter
(seed and circumcenter remain on the same side of the edge). Added a
distance cap: circumcenters beyond 5×cellSize from the edge midpoint
are discarded as degenerate. Reduces max segment from 16 778 to 669
blocks, average from 201 to 117.

+++ app/webApp/src/wasmJsMain/kotlin/org/micoli/micraft/LocalPlayerController.kt
@@ -682,7 +682,7 @@ class LocalPlayerController(
             val debugCz = hudZ.toInt().floorDiv(WorldConstants.CHUNK_SIZE)
             jsUpdateChunkDebug(
                 chunkManager.getChunkDebugJson(
-                    debugCx, debugCz, WorldConstants.FORWARD_VIEW_RADIUS, yaw.toDouble()))
+                    debugCx, debugCz, WorldConstants.FORWARD_VIEW_RADIUS, yaw))
         }
     }
